requirements.txt
Django~=5.0.6
djangorestframework~=3.15.2
assemblyai~=0.34.0
nltk~=3.8.1